* TableViewDataSource.m some NSInteger conversions
git-svn-id: svn://svn.savannah.nongnu.org/gap/trunk@2217 d25b4930-3292-e311-b074-0010c6de5da8

+++ system-apps/Zipper/TableViewDataSource.m
@@ -39,13 +39,13 @@
 //------------------------------------------------------------------------------
 // Implementation NSTableView DataSource
 //------------------------------------------------------------------------------
-- (int)numberOfRowsInTableView:(NSTableView *)aTableView
+- (NSInteger)numberOfRowsInTableView:(NSTableView *)aTableView
 {
     return [_archive elementCount];
 }
 
 - (void) tableView: (NSTableView *)tableView willDisplayCell: (id)aCell
-   forTableColumn: (NSTableColumn *)tableColumn row: (int)row
+   forTableColumn: (NSTableColumn *)tableColumn row: (NSInteger)row
 {
 	NSImage *image;
 	FileInfo *fileInfo = [_archive elementAtIndex: row];
@@ -63,7 +63,7 @@
 
 
 - (id)tableView:(NSTableView *)aTableView objectValueForTableColumn:(NSTableColumn *)aTableColumn
-    row:(int)rowIndex
+    row:(NSInteger)rowIndex
 {
 	FileInfo *fileInfo = [_archive elementAtIndex:rowIndex];
